requirements.txt
attrs==24.2.0
certifi==2022.12.7
charset-normalizer==3.4.0
click==8.1.7
future==0.18.3
h11==0.14.0
idna==3.10
jaraco.classes==3.4.0
keyring==23.13.1
keyrings.alt==4.2.0
more-itertools==10.5.0
outcome==1.3.0.post0
pillow==11.0.0
pyicloud==1.0.0
PySocks==1.7.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
